speed up constructors
It turns out that `goto &somefunc;` is slower than just doing a plain
naked @_-reusing `&somefunc;` call – even though that incurs a double
return. The upshot is that the goto form should only be used if the
called function is sensitive to the call stack and needs the current
frame cleared out – which is not the case here.